requirements.txt
numpy<2
scipy==1.7.2 #changed this since it was causing issues with the latest version of scipy
pytest==7.1.3
pytest-cov==4.0.0
scikit-learn
pillow
matplotlib==3.5.0
librosa==0.9.2
pyomyo==0.0.5
pyserial
wfdb
bleak
requests
h5py
# For Docs
sphinx==5.0.0
sphinx_rtd_theme==1.0.0
myst_parser==0.18.1
linkify-it-py==2.0.0
# For testing
pytest-skip-slow==0.0.3
dearpygui
# For oymotion - only on linx
#bluepy
#PyWavelets==1.4.1
# for new sgt
dearpygui
opencv-python
datetime
websockets==8.1
h5py
onedrivedownloader
sifi-bridge-py